this changes everything to use the flex&bison parser again.

+++ include/ad9910.h
@@ -224,7 +224,7 @@ INLINE void ad9910_update_profile_reg(uint8_t profile);
  * then call ad9910_update_reg on that register directly */
 INLINE void ad9910_update_matching_reg(ad9910_register_bit field);
 
-uint32_t ad9910_convert_frequency(double f);
+uint32_t ad9910_convert_frequency(float f);
 
 void ad9910_init(void);
