+++ src/ad9910.c
@@ -188,7 +188,7 @@ ad9910_enable_parallel(int mode)
 }
 
 uint32_t
-ad9910_convert_frequency(double f)
+ad9910_convert_frequency(float f)
 {
   return nearbyint(f / 1e9 * 0xFFFFFFFF);
 }
